requirements/dev.txt
-r basic.txt

# Debug toolbar
Flask-DebugToolbar==0.10.1
